+++ pkg/analyzer/lib/src/clients/dart_style/rewrite_cascade.dart
@@ -36,24 +36,27 @@ ExpressionStatement fixCascadeByParenthesizingTarget({
 
 /// Recursively insert [cascadeTarget] (the LHS of the cascade) into the
 /// LHS of the assignment expression that used to be the cascade's RHS.
-Expression insertCascadeTargetIntoExpression({
+ExpressionImpl insertCascadeTargetIntoExpression({
   required Expression expression,
   required Expression cascadeTarget,
 }) {
+  expression as ExpressionImpl;
+  cascadeTarget as ExpressionImpl;
+
   // Base case: We've recursed as deep as possible.
   if (expression == cascadeTarget) return cascadeTarget;
 
   // Otherwise, copy `expression` and recurse into its LHS.
-  if (expression is AssignmentExpression) {
+  if (expression is AssignmentExpressionImpl) {
     return AssignmentExpressionImpl(
       leftHandSide: insertCascadeTargetIntoExpression(
         expression: expression.leftHandSide,
         cascadeTarget: cascadeTarget,
-      ) as ExpressionImpl,
+      ),
       operator: expression.operator,
-      rightHandSide: expression.rightHandSide as ExpressionImpl,
+      rightHandSide: expression.rightHandSide,
     );
-  } else if (expression is IndexExpression) {
+  } else if (expression is IndexExpressionImpl) {
     var expressionTarget = expression.realTarget;
     var question = expression.question;
 
@@ -63,7 +66,7 @@ Expression insertCascadeTargetIntoExpression({
       question = _synthesizeToken(TokenType.QUESTION, expression.period!);
     }
 
-    return astFactory.indexExpressionForTarget2(
+    return IndexExpressionImpl.forTarget(
       target: insertCascadeTargetIntoExpression(
         expression: expressionTarget,
         cascadeTarget: cascadeTarget,
@@ -79,7 +82,7 @@ Expression insertCascadeTargetIntoExpression({
       target: insertCascadeTargetIntoExpression(
         expression: expressionTarget,
         cascadeTarget: cascadeTarget,
-      ) as ExpressionImpl,
+      ),
       // If we've reached the end, replace the `..` operator with `.`
       operator: expressionTarget == cascadeTarget
           ? _synthesizeToken(TokenType.PERIOD, expression.operator!)
@@ -88,7 +91,7 @@ Expression insertCascadeTargetIntoExpression({
       typeArguments: expression.typeArguments,
       argumentList: expression.argumentList,
     );
-  } else if (expression is PropertyAccess) {
+  } else if (expression is PropertyAccessImpl) {
     var expressionTarget = expression.realTarget;
     return astFactory.propertyAccess(
       insertCascadeTargetIntoExpression(